Make `bsp_read_matrix` statically linked instead of header-only.
